Part of optimize and xml file
