+++ .gitignore
@@ -3,3 +3,4 @@ output.txt
 input.txt
 __pycache__
 *.lp
+.DS_Store
