+++ seir-refactor/paramsT1.py
@@ -1,9 +1,9 @@
 # Run parameters for seir_opt script
 
-verbosity = 2           # 0 least verbose output ... 3 most verbose 
+verbosity = 3           # 0 least verbose output ... 3 most verbose 
 simulate_only = True
 realloc_flag = False    # proportional reallocation of unused vacc to non-donors in sim
-# non_donor_deaths = False # include non-donor deaths in objective 
+non_donor_deaths = False # include non-donor deaths in objective 
 p_k = 0.75       # maximum proportion of available vacc used in donor area
 
 lambda_0 = 1    # initial Lagrange multiplier for infection
